+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [6.1.1](https://github.com/medic/cht-conf/compare/v6.1.0...v6.1.1) (2026-04-10)
+
+
+### Bug Fixes
+
+* **#799:** batch contact IDs in move-contacts to avoid URI length errors ([#802](https://github.com/medic/cht-conf/issues/802)) ([08e92a5](https://github.com/medic/cht-conf/commit/08e92a568196bdcb3e68076a760fbd8a5b7ff461)), closes [#799](https://github.com/medic/cht-conf/issues/799) [#799](https://github.com/medic/cht-conf/issues/799) [#799](https://github.com/medic/cht-conf/issues/799) [#799](https://github.com/medic/cht-conf/issues/799) [#799](https://github.com/medic/cht-conf/issues/799) [#799](https://github.com/medic/cht-conf/issues/799) [#799](https://github.com/medic/cht-conf/issues/799) [#799](https://github.com/medic/cht-conf/issues/799) [#799](https://github.com/medic/cht-conf/issues/799) [#799](https://github.com/medic/cht-conf/issues/799) [#799](https://github.com/medic/cht-conf/issues/799) [#799](https://github.com/medic/cht-conf/issues/799) [#799](https://github.com/medic/cht-conf/issues/799) [#799](https://github.com/medic/cht-conf/issues/799) [#799](https://github.com/medic/cht-conf/issues/799) [#799](https://github.com/medic/cht-conf/issues/799)
+
+# [6.1.0](https://github.com/medic/cht-conf/compare/v6.0.2...v6.1.0) (2026-04-02)
+
+
+### Features
+
+* **#653:** add --merge-primary-contacts flag for merge-contacts action ([#654](https://github.com/medic/cht-conf/issues/654)) ([def533d](https://github.com/medic/cht-conf/commit/def533dbca20abca8303aea7f314921c6a1c166b)), closes [#653](https://github.com/medic/cht-conf/issues/653)
+
 ## [6.0.2](https://github.com/medic/cht-conf/compare/v6.0.1...v6.0.2) (2026-01-16)